clean-up test code. (#188)
