+++ src/test/java/com/odysseusinc/athena/service/concept/SolrConceptSearchAbstractTest.java
@@ -7,13 +7,8 @@ import com.odysseusinc.athena.api.v1.controller.dto.ConceptSearchDTO;
 import com.odysseusinc.athena.service.impl.ConceptSearchPhraseToSolrQueryService;
 import com.odysseusinc.athena.service.impl.ConceptSearchQueryPartCreator;
 import com.odysseusinc.athena.service.impl.QueryDebugUtils;
-import java.io.File;
 import java.io.IOException;
-import java.nio.charset.Charset;
 import java.util.Collections;
-import java.util.Map;
-import org.apache.commons.io.FileUtils;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.solr.client.solrj.SolrQuery;
 import org.apache.solr.client.solrj.SolrServerException;
 import org.apache.solr.client.solrj.response.QueryResponse;
@@ -57,37 +52,17 @@ public abstract class SolrConceptSearchAbstractTest {
         ConceptSearchDTO conceptSearchDTO = createConceptSearchDTO(queryString);
         SolrQuery query = conceptSearchDTOToSolrQuery.createQuery(conceptSearchDTO, Collections.emptyList());
 
-        QueryResponse queryResponse = debug(testName.getMethodName(), query,
-                () -> SolrInitializer.server.query(query));
-        return queryResponse.getResults();
-    }
-
-
-    private QueryResponse debug(String dirName, SolrQuery query, ProducerWithException<QueryResponse> requestFunction) throws SolrServerException, IOException {
+        QueryResponse queryResponse;
         if (!DEBUG_INFO_ENABLED) {
-            return requestFunction.produce();
+            queryResponse = SolrInitializer.server.query(query);
+        } else {
+            queryResponse = QueryDebugUtils.debug(
+                    DEBUG_INFO_FOR_SEARCH_DIR_BASE, testName.getMethodName(),
+                    query, () -> SolrInitializer.server.query(query)
+            );
         }
 
-        QueryDebugUtils.addDebugAndScore(query);
-        QueryResponse response = requestFunction.produce();
-
-        String dir = String.format("%s/%s/", DEBUG_INFO_FOR_SEARCH_DIR_BASE, dirName);
-        FileUtils.deleteDirectory(new File(dir));
-
-        File fileWithQuery = new File(String.format("%s%s", dir,  "query"));
-        FileUtils.writeStringToFile(fileWithQuery, QueryDebugUtils.getQuery(query.getQuery()), Charset.defaultCharset());
-
-        for (Map.Entry<String, String> stringStringEntry : response.getExplainMap().entrySet()) {
-            String k = stringStringEntry.getKey();
-            String v = stringStringEntry.getValue();
-            File file = new File(dir + StringUtils.substringBetween(v, "\n", "=") + "_" + k);
-            FileUtils.writeStringToFile(file, v, Charset.defaultCharset());
-        }
-        return response;
-    }
-
-    public interface ProducerWithException<T> {
-        T produce() throws SolrServerException, IOException;
+        return queryResponse.getResults();
     }
 
 }
